fixed status and added test

+++ src/test_repo.rb
@@ -11,6 +11,7 @@ class TestRepository < Minitest::Test
   def setup
     
   end
+  
   # check to see if framework is working
   def test_simple
     assert_equal(4, 2+2)
@@ -26,14 +27,14 @@ class TestRepository < Minitest::Test
     Dir.chdir('.test')
     Repository.create()
     assert(File.exist?('.repository'))
-    assert(File.exist?('.repository/.files'))
     assert(File.exist?('.repository/.stage'))
-    FileUtils.rm_rf('.test')
     Dir.chdir('..')
+    FileUtils.rm_rf('.test')
   end
 
   #----------------------------------------------------------------------
   def test_status
+    # tests blank status
     if File.exist?('.test')
       FileUtils.rm_rf('.test')
     end
@@ -41,8 +42,23 @@ class TestRepository < Minitest::Test
     Dir.chdir('.test')
     Repository.create()
     Repository.status()
-    FileUtils.rm_rf('.test')
     Dir.chdir('..')
+    FileUtils.rm_rf('.test')
+  end
+
+  #----------------------------------------------------------------------
+  def test_status2
+    # adds a file and checks if in status
+    if File.exist?('.test')
+      FileUtils.rm_rf('.test')
+    end
+    Dir.mkdir('.test')
+    Dir.chdir('.test')
+    Repository.create()
+    File.open('file1.txt', 'w') { |f| f.write("test text") }
+    Repository.status()
+    Dir.chdir('..')
+    FileUtils.rm_rf('.test')
   end
 
   #----------------------------------------------------------------------
@@ -57,8 +73,8 @@ class TestRepository < Minitest::Test
     a = ['.test_file1']
     Repository.add(a)      
     Repository.status()
-    #FileUtils.rm_rf('.test')
     Dir.chdir('..')
+    FileUtils.rm_rf('.test')
   end
 
   #----------------------------------------------------------------------
